+++ CHANGELOG.md
@@ -13,10 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `opentelemetry-instrumentation-sqlalchemy` Update unit tests to run with SQLALchemy 2
+  ([#2976](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2976))
+
 ### Fixed
 
 ### Breaking changes
 
+- `opentelemetry-instrumentation-sqlalchemy` teach instruments version
+  ([#2971](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2971))
 - Drop `opentelemetry-instrumentation-test` package from default instrumentation list
   ([#2969](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2969))